pyproject.toml
[tool.poetry]
name = "ml_tech_challenge"
version = "0.1.0"
description = "Mercado Libre technical challenge"
authors = ["Leonardo Darre <lrdarre@gmail.com>"]

[tool.poetry.dependencies]
python = "~3.10"
pandas = "^1.5.0"
seaborn = "^0.12.0"
jupyter = "^1.0.0"
ipyplot = "^1.1.0"
xgboost = "~1.5.2"
sklearn = "^0.0"
category-encoders = "^2.2.2"
plotly = "^5.6.0"
umap-learn = "^0.5.2"
hdbscan = "^0.8.28"
sentence-transformers = "^2.2.0"
bokeh = "~2.4.2"
datashader = "^0.13.0"
holoviews = "^1.14.8"
scikit-image = "^0.19.2"
colorcet = "^3.0.0"
faiss-cpu = "^1.7.2"

# Formatting
black = "22.6.0"
isort = "^5.9.3"
flake8 = "^3.9.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
